refactor(core): plugin supervisor consumes sources, config side owns discovery and watching
