+++ packages/core/src/config/plugin/source.ts
@@ -1,0 +1,175 @@
+export * as ConfigPluginSource from "./source"
+
+import { Directory, Document, type Entry } from "@opencode-ai/schema/config"
+import { ConfigPlugin } from "@opencode-ai/schema/config/plugin"
+import { FSUtil } from "@opencode-ai/util/fs-util"
+import { LayerNode } from "@opencode-ai/util/effect/layer-node"
+import { Context, Effect, Layer, Option, PubSub, Scope, Stream } from "effect"
+import path from "path"
+import { fileURLToPath } from "url"
+import { Config } from "../../config"
+import { Watcher } from "../../filesystem/watcher"
+import { Location } from "../../location"
+
+export type Operation =
+  | {
+      readonly type: "add"
+      readonly target: string
+      readonly options: Record<string, unknown>
+      readonly mtime?: number
+    }
+  | {
+      readonly type: "remove"
+      readonly target: string
+    }
+
+export interface Interface {
+  readonly operations: () => Effect.Effect<readonly Operation[], never, Scope.Scope>
+  readonly changes: () => Stream.Stream<void>
+}
+
+export class Service extends Context.Service<Service, Interface>()("@opencode/ConfigPluginSource") {}
+
+export type Options = {
+  readonly dynamic?: boolean
+}
+
+export const layer = (options?: Options) =>
+  Layer.effect(
+    Service,
+    Effect.gen(function* () {
+      if (options?.dynamic === false) {
+        return Service.of({
+          operations: () => Effect.succeed([]),
+          changes: () => Stream.empty,
+        })
+      }
+
+      const config = yield* Config.Service
+      const watcher = yield* Watcher.Service
+      const fs = yield* FSUtil.Service
+      const location = yield* Location.Service
+      const configuredChanges = yield* PubSub.unbounded<void>()
+      const watched = new Set<string>()
+
+      // Configured local plugin files can live outside config roots, where the
+      // config change feed cannot see them; watch those entrypoints directly.
+      // Watches start on first sighting and are never torn down individually:
+      // a stale watch after a config edit costs one deduped fs handle and a
+      // no-op activation, and every watch dies with this layer's scope.
+      const watchConfiguredSources = Effect.fn("ConfigPluginSource.watchConfiguredSources")(function* (
+        entries: readonly Entry[],
+        operations: readonly Operation[],
+      ) {
+        for (const operation of operations) {
+          if (operation.type !== "add" || !path.isAbsolute(operation.target)) continue
+          if (watched.has(operation.target)) continue
+          // The config change feed already covers {plugin,plugins} directories.
+          if (isPluginSource(entries, operation.target)) continue
+          // Directory targets can't hot-reload (their stat mtime ignores edits
+          // inside), so don't watch what can't trigger anything.
+          if (yield* fs.isDir(operation.target)) continue
+          watched.add(operation.target)
+          const updates = yield* watcher.subscribe({ path: operation.target, type: "file" })
+          yield* updates.pipe(
+            Stream.runForEach(() => PubSub.publish(configuredChanges, undefined)),
+            Effect.catchCause((cause) =>
+              Effect.logError("configured plugin watch failed", { target: operation.target, cause }),
+            ),
+            Effect.forkScoped({ startImmediately: true }),
+          )
+        }
+      })
+
+      return Service.of({
+        operations: Effect.fn("ConfigPluginSource.operations")(function* () {
+          const entries = yield* config.entries()
+          const operations = yield* scan(fs, location, entries)
+          yield* watchConfiguredSources(entries, operations)
+          return operations
+        }),
+        changes: () =>
+          Stream.merge(
+            config.changes().pipe(
+              Stream.filterEffect((update) =>
+                Effect.map(config.entries(), (entries) => isPluginSource(entries, update.path)),
+              ),
+              Stream.map(() => undefined),
+            ),
+            Stream.fromPubSub(configuredChanges),
+          ),
+      })
+    }),
+  )
+
+export const requirements = LayerNode.group([Config.node, FSUtil.node, Location.node, Watcher.node])
+
+function parse(input: ConfigPlugin.Plugin): Operation {
+  if (typeof input !== "string") {
+    return { type: "add", target: input.package, options: input.options ?? {} }
+  }
+  if (!input.startsWith("-")) return { type: "add", target: input, options: {} }
+  if (input.length === 1) throw new Error("Plugin remove operation requires a target")
+  return { type: "remove", target: input.slice(1) }
+}
+
+const scan = Effect.fn("ConfigPluginSource.scan")(function* (
+  fs: FSUtil.Interface,
+  location: Location.Interface,
+  entries: readonly Entry[],
+) {
+  const discovered = yield* Effect.forEach(
+    entries.filter((entry): entry is Directory => entry.type === "directory"),
+    (entry) => discoverDirectory(fs, entry.path),
+  ).pipe(Effect.map((items) => items.flat()))
+  const configured = entries
+    .filter((entry): entry is Document => entry.type === "document")
+    .flatMap((entry) =>
+      (entry.info.plugins ?? []).map(parse).map((operation) => {
+        if (operation.type === "remove") return operation
+        const directory = entry.path ? path.dirname(entry.path) : location.directory
+        const target = operation.target.startsWith("file://")
+          ? fileURLToPath(operation.target)
+          : operation.target.startsWith("./") || operation.target.startsWith("../")
+            ? path.resolve(directory, operation.target)
+            : operation.target
+        return { ...operation, target }
+      }),
+    )
+  // Explicit config is applied last so it can remove auto-discovered packages.
+  return yield* Effect.forEach([...discovered, ...configured], (operation) => {
+    if (operation.type === "remove" || !path.isAbsolute(operation.target)) return Effect.succeed(operation)
+    return fs.stat(operation.target).pipe(
+      Effect.map((info) => ({
+        ...operation,
+        mtime: Option.getOrElse(info.mtime, () => new Date(0)).getTime(),
+      })),
+      Effect.catch(() => Effect.succeed(operation)),
+    )
+  })
+})
+
+function discoverDirectory(fs: FSUtil.Interface, directory: string) {
+  return Effect.gen(function* () {
+    const files = yield* fs
+      .scan("{plugin,plugins}/*.{ts,js}", {
+        cwd: directory,
+        absolute: true,
+        include: "file",
+        dot: true,
+        symlink: true,
+      })
+      .pipe(Effect.orElseSucceed(() => []))
+    return files.sort().map((target): Operation => ({ type: "add", target, options: {} }))
+  })
+}
+
+const sourceDirectories = ["plugin", "plugins"] as const
+
+function isPluginSource(entries: readonly Entry[], file: string) {
+  return entries.some(
+    (entry) =>
+      entry.type === "directory" &&
+      sourceDirectories.some((directory) => FSUtil.contains(path.join(entry.path, directory), file)),
+  )
+}
